% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_details.R
\name{drug_details}
\alias{drug_details}
\title{Search for the official name and code of BNF sections, chemicals and presentations.}
\usage{
drug_details(name = NULL, chemical_section_or_presentation_code = NULL,
  BNF_section = NULL, exact_name_or_code = NULL)
}
\arguments{
\item{name}{Drug name.}

\item{chemical_section_or_presentation_code}{An ID unique to BNF sections, chemicals or presentations.}

\item{BNF_section}{The BNF section.}

\item{exact_name_or_code}{Search by an exact name/code(ID).}
}
\value{
Returns official names and codes(ID) of BNF sections, chemicals and presentations.
}
\description{
Search for the official name and code of BNF sections, chemicals and presentations.
}
\examples{
All BNF sections, chemicals and presentations matching a name (case-insensitive) = drug_details(name= "lipid")
All chemicals and presentations matching a code = drug_details(chemical_section_or_presentation_code= "0212000AA")
All BNF sections matching a code = drug_details(BNF_section= "2.4")
All BNF sections, chemicals and presentations exactly matching a name or code = drug_details(exact_name_or_code= "212000AA")
Or a variation of the above (with at least 1 name, chemical_section_or_presentation_code, BNF_section, or exact_name_or_code).
N.B For now the use of multiple terms in one category requires "&q=", e.g drug_details(name="lipid&q=drug")
Read the [wiki](https://github.com/fergustaylor/openprescribingR/wiki) for more help.
}
