% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_function.R
\name{location_function}
\alias{location_function}
\title{Search for the locations of all (GP) practices, by code. Returns GeoJSON or sfc.}
\usage{
location_function(CCG_code, as_sf = TRUE/FALSE)
}
\arguments{
\item{CCG_code}{A code assigned to each CCG, found using organisation_details().}

\item{as_sf}{If TRUE, Returns as Simple Features Collection instead.}
}
\value{
Returns the locations of all GP practices in a CCG). Returns GeoJSON (or sfc).
}
\description{
Search for the locations of all (GP) practices, by code. Returns GeoJSON or sfc.
}
\examples{
Location (approximate) of all practices in a CCG = location_function(CCG_code, as_sf = TRUE/FALSE)
location_function("99H", as_sf = TRUE)
Read the [wiki](https://github.com/fergustaylor/openprescribingR/wiki) for more help.
}
