% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_by_code.R
\name{spending_by_code}
\alias{spending_by_code}
\title{Queries from August 2010 to date and returns total spending and items by month.}
\usage{
spending_by_code(BNF_code = NULL)
}
\arguments{
\item{BNF_code}{A code found using drug_details(). Or the BNF.}
}
\value{
Returns total spending and items by month.
}
\description{
Queries from August 2010 to date and returns total spending and items by month.
}
\examples{
spending_by_code() 
spending_by_code(BNF_code = "0212")
}
