% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_by_practice.R
\name{spending_by_practice}
\alias{spending_by_practice}
\title{Queries from August 2010 to date and returns total spending and items by practice by month.
You must specify either an organisation, or a date.}
\usage{
spending_by_practice(BNF_section_code = NULL, chemical_code = NULL,
  presentation_code = NULL, practice_code = NULL, CCG_code = NULL,
  date_code = NULL)
}
\arguments{
\item{BNF_section_code}{A code found using drug_details().}

\item{chemical_code}{A code found using drug_details().}

\item{presentation_code}{A code found using drug_details().}

\item{practice_code}{A code found using organisation_codes().}

\item{CCG_code}{A code found using organisation_codes().}

\item{date_code}{A date using the format "2015-12-01".}
}
\value{
Returns total spending and items by practice by month.
}
\description{
Queries from August 2010 to date and returns total spending and items by practice by month.
You must specify either an organisation, or a date.
}
\examples{
spending_by_practice(BNF_section_code = "0212", date_code = "2015-04-01")
spending_by_practice(chemical_code = "0212000AA", date_code = "2015-04-01")
spending_by_practice(practice_code = "H81068")
spending_by_practice(CCG_code = "03V")
Or a variation of the above (with at least 1 practice_code, date_code, or CCG_code).
}
