% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organisation_codes.R
\name{organisation_codes}
\alias{organisation_codes}
\title{Search for details about a CCG or practice by code or name.}
\usage{
organisation_codes(organisation_code_or_name = NULL,
  CCG_code_or_name = NULL, practice_code_or_name = NULL,
  exact_name_or_code = NULL)
}
\arguments{
\item{organisation_code_or_name}{A code found using drug_details().}

\item{CCG_code_or_name}{A code found using drug_details().}

\item{practice_code_or_name}{A code found using drug_details().}

\item{exact_name_or_code}{A code found using drug_details().}
}
\value{
Returns details about a CCG or practice by code or name.
}
\description{
Search for details about a CCG or practice by code or name.
}
\examples{
All organisations matching a code or name = organisation_codes(organisation_code_or_name= "...")
All CCGs matching a code or name = organisation_codes(CCG_code_or_name= "...")
#' All practices matching a code or name = organisation_codes(practice_code_or_name= "...")
All organisations exactly matching a code or name = organisation_codes(exact_name_or_code= "...")
Or a variation of the above (with at least 1 organisation_code_or_name, CCG_code_or_name, practice_code_or_name, or exact_name_or_code).
N.B For now the use of multiple terms in one category requires "&q=", e.g organisation_codes(organisation_code_or_name= "Beaumont&q=Gloucester")
}
